backend/requirements/base.txt
# ============================================================================
# Core Application Dependencies
# ============================================================================
# These are the essential packages required for the application to run
# in any environment (development, staging, production)

# Web Framework and ASGI Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database - PostgreSQL with async support
sqlalchemy[asyncio]==2.0.25  # SQLAlchemy ORM with async support
asyncpg==0.29.0  # PostgreSQL async driver for FastAPI
psycopg2-binary==2.9.9  # PostgreSQL sync driver (for Alembic migrations)
pgvector==0.2.5  # Vector extension for RAG/AI features

# Data Validation and Settings
pydantic==2.4.2
pydantic-settings==2.0.3
email-validator==2.2.0

# Task Queue and Background Jobs
celery==5.3.4
redis==5.0.1

# AI Provider Integration
openai>=1.57.0
httpx==0.27.0  # Modern HTTP client for API calls

# RAG (Retrieval Augmented Generation) - Phase 11
langchain==0.1.0  # LangChain framework for RAG orchestration
langchain-openai==0.0.3  # OpenAI integration for LangChain
langchain-community==0.0.13  # Community integrations
tiktoken==0.5.2  # Token counting for embeddings

# Authentication and Security
python-jose[cryptography]==3.3.0  # JWT token generation
PyJWT==2.9.0
passlib[bcrypt]==1.7.4  # Password hashing
bcrypt==4.0.1
python-magic==0.4.27  # MIME type detection for file validation
bleach>=6.0.0  # HTML sanitization for XSS prevention
# python-magic-bin==0.4.14  # Windows-only binary (not needed in Linux containers)

# File Processing
pdf2image==1.17.0  # PDF to image conversion
Pillow==10.4.0  # Image processing
PyPDF2==3.0.1  # PDF manipulation

# Data Export and Processing
pandas==2.1.3  # Data manipulation
openpyxl==3.1.2  # Excel file support
reportlab==4.0.7  # PDF generation

# Analytics and Performance
duckdb==1.1.0  # Embedded analytical database
pyarrow==14.0.1  # Columnar data format

# Cloud Storage
boto3==1.34.0  # AWS S3 and compatible storage

# Async I/O and WebSockets
aiofiles==23.2.1  # Async file operations
websockets==12.0  # WebSocket support

# HTTP Client
requests==2.32.3  # HTTP library for external APIs

# Web Scraping (for website data sources)
beautifulsoup4==4.12.3  # HTML parsing for website crawling
lxml==5.3.0  # Fast HTML/XML parser backend

# Environment Variables
python-dotenv==1.0.0  # Load .env files
